stop deleting block when typing backspace on editor

+++ src/viewport.js
@@ -488,7 +488,11 @@ function onkeydown(state, event) {
     case "Backspace":
       // Only handle block deletion if not in input field, a block is selected, and not in edit mode
       const selectedBlockId = getFirstSelectedBlockId(state);
-      if (selectedBlockId !== null && currentPage.editingId === null) {
+      if (
+        selectedBlockId !== null &&
+        currentPage.editingId === null &&
+        !currentPage.isEditorFocused
+      ) {
         event.preventDefault();
         return deleteSelectedBlocks(state);
       }
